+++ packages/platform-server/plugins/template-checker.ts
@@ -9,15 +9,15 @@ import { IComponent, InDiv } from '@indiv/core';
  * @param {IComponent} component
  */
 export function templateChecker(component: IComponent) {
-  if (component.template) return;
-  if (component.templateUrl) {
-    let templatePath = component.templateUrl;
-    const templateRootPath = (component.indivInstance as InDiv).getTemplateRootPath;
+  if (component.$template) return;
+  if (component.$templateUrl) {
+    let templatePath = component.$templateUrl;
+    const templateRootPath = component.$indivInstance.getTemplateRootPath;
     if (templateRootPath) {
-      if (path.isAbsolute(component.templateUrl)) templatePath = `${templateRootPath}${component.templateUrl}`;
-      else templatePath = path.resolve(templateRootPath, component.templateUrl);
+      if (path.isAbsolute(component.$templateUrl)) templatePath = `${templateRootPath}${component.$templateUrl}`;
+      else templatePath = path.resolve(templateRootPath, component.$templateUrl);
     }
     const templateString = fs.readFileSync(templatePath).toString();
-    component.template = templateString;
+    component.$template = templateString;
   }
 }